Adjust comeback score

+++ win_prediction_model.py
@@ -77,52 +77,60 @@ def get_game_play_by_play(game):
     # Get play by play data
     pbp = playbyplay.PlayByPlay(game["gameId"]).get_data_frames()[0]
 
-    # Get desired data for model training. Setting up columns GAME_ID, AWAY_SCORE, HOME_SCORE, SCORE_MARGIN, HOME_WIN, OT 
+    # Get desired data for model training. Setting up columns GAME_ID, AWAY_SCORE, HOME_SCORE, SCORE_MARGIN, HOME_WIN, OT
     res = pd.DataFrame()
-    res['GAME_ID'] = pbp["GAME_ID"].astype(str)
-    res['SCORE'] = pbp['SCORE'].copy()
+    res["GAME_ID"] = pbp["GAME_ID"].astype(str)
+    res["SCORE"] = pbp["SCORE"].copy()
 
     # Edit SCORE to eliminate empty scores
-    prev_score = '0 - 0'
+    prev_score = "0 - 0"
     for i in range(len(res)):
         play = res.iloc[[i]]
-        if res.loc[i,'SCORE'] == None:
-            res.loc[i,'SCORE'] = prev_score
+        if res.loc[i, "SCORE"] == None:
+            res.loc[i, "SCORE"] = prev_score
         else:
-            prev_score = res.loc[i,'SCORE']
+            prev_score = res.loc[i, "SCORE"]
 
     # Split SCORE into AWAY_SCORE and HOME_SCORE
-    res[['AWAY_SCORE', 'HOME_SCORE']] = res['SCORE'].str.split('-', expand=True).astype(int)
-    res.drop('SCORE', axis=1, inplace=True)
+    res[["AWAY_SCORE", "HOME_SCORE"]] = (
+        res["SCORE"].str.split("-", expand=True).astype(int)
+    )
+    res.drop("SCORE", axis=1, inplace=True)
 
     # Set up SCORE_MARGIN to remove 'TIE' and None values
-    res['SCORE_MARGIN'] = pbp['SCOREMARGIN'].copy()
-    prev_margin = '0'
+    res["SCORE_MARGIN"] = pbp["SCOREMARGIN"].copy()
+    prev_margin = "0"
     for i in range(len(res)):
         play = res.iloc[[i]]
-        if res.loc[i,'SCORE_MARGIN'] == None:
-            res.loc[i,'SCORE_MARGIN'] = prev_margin
-        elif res.loc[i,'SCORE_MARGIN'] == 'TIE':
-            res.loc[i,'SCORE_MARGIN'] = '0'
-            prev_margin = '0'
+        if res.loc[i, "SCORE_MARGIN"] == None:
+            res.loc[i, "SCORE_MARGIN"] = prev_margin
+        elif res.loc[i, "SCORE_MARGIN"] == "TIE":
+            res.loc[i, "SCORE_MARGIN"] = "0"
+            prev_margin = "0"
         else:
-            prev_margin = res.loc[i,'SCORE_MARGIN']
-    res['SCORE_MARGIN'] = res['SCORE_MARGIN'].astype(int)
+            prev_margin = res.loc[i, "SCORE_MARGIN"]
+    res["SCORE_MARGIN"] = res["SCORE_MARGIN"].astype(int)
 
     # Set up HOME_WIN based on SCOREMARGIN
-    res['HOME_WIN'] = 1 if int(pbp.tail(1)['SCOREMARGIN'].values[0]) > 0 else 0
-    res['OT'] = 0
+    res["HOME_WIN"] = 1 if int(pbp.tail(1)["SCOREMARGIN"].values[0]) > 0 else 0
+    res["OT"] = 0
 
     # lambda function that converts the PCTIMESTRING to seconds using datetime
-    time_to_seconds = lambda x: datetime.datetime.strptime(x, '%M:%S').minute * 60 + datetime.datetime.strptime(x, '%M:%S').second
-    res['TIME_REMAINING'] = ((4 - pbp['PERIOD']) *12*60) + pbp['PCTIMESTRING'].apply(time_to_seconds)
+    time_to_seconds = (
+        lambda x: datetime.datetime.strptime(x, "%M:%S").minute * 60
+        + datetime.datetime.strptime(x, "%M:%S").second
+    )
+    res["TIME_REMAINING"] = ((4 - pbp["PERIOD"]) * 12 * 60) + pbp["PCTIMESTRING"].apply(
+        time_to_seconds
+    )
 
     # Handling time remaining for overtime
-    res.loc[res['TIME_REMAINING'] < 0, 'OT'] = 1
-    res.loc[res['TIME_REMAINING'] < 0, 'TIME_REMAINING'] = pbp['PCTIMESTRING'].apply(time_to_seconds)
+    res.loc[res["TIME_REMAINING"] < 0, "OT"] = 1
+    res.loc[res["TIME_REMAINING"] < 0, "TIME_REMAINING"] = pbp["PCTIMESTRING"].apply(
+        time_to_seconds
+    )
 
     return res
-
 
 
 def get_comeback_score(game):
@@ -134,7 +142,9 @@ def get_comeback_score(game):
 
     # Run model on the game to obtain win probabilities at each play
     model_results = run_model(
-        play_by_play[["TIME_REMAINING", "HOME_SCORE", "AWAY_SCORE", "SCORE_MARGIN", "OT"]]
+        play_by_play[
+            ["TIME_REMAINING", "HOME_SCORE", "AWAY_SCORE", "SCORE_MARGIN", "OT"]
+        ]
     )
 
     # Make predictions more coherant by adding 100% win probablity at the end of the game to winning team. Calculate lowest win probably of the winning team
@@ -146,7 +156,7 @@ def get_comeback_score(game):
     else:
         predictions = np.append(model_results, [[1, 0]], axis=0)
         for p in predictions:
-            lowest_win_prob = min(lowest_win_prob, p[0])   
+            lowest_win_prob = min(lowest_win_prob, p[0])
 
     # plt.plot(predictions)
     # plt.show()
